# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Golden Years API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "https://api.goldenyears.local",
      "description": "Generated contract placeholder."
    }
  ],
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "validation_failed"
              },
              "message": {
                "type": "string",
                "example": "Request validation failed."
              },
              "details": {
                "type": "object",
                "additionalProperties": {}
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
      "EmptyJsonBody": {
        "type": "object",
        "properties": {}
      },
      "user_auth_loginResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UserLoginResponseData"
          }
        },
        "required": [
          "data"
        ]
      },
      "UserLoginResponseData": {
        "type": "object",
        "properties": {
          "session": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "audience": {
                "type": "string",
                "enum": [
                  "user"
                ]
              },
              "expires_at": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "audience",
              "expires_at"
            ],
            "additionalProperties": false
          },
          "user": {
            "$ref": "#/components/schemas/SafeUser"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "csrf_token": {
            "type": "string"
          }
        },
        "required": [
          "session",
          "user",
          "roles",
          "csrf_token"
        ],
        "additionalProperties": false
      },
      "SafeUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "display_name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "display_name"
        ],
        "additionalProperties": false
      },
      "UserLoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "email",
          "password"
        ],
        "additionalProperties": false
      },
      "create_sessionResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UserLoginResponseData"
          }
        },
        "required": [
          "data"
        ]
      },
      "user_auth_logoutResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UserLogoutResponseData"
          }
        },
        "required": [
          "data"
        ]
      },
      "UserLogoutResponseData": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "additionalProperties": false
      },
      "delete_sessionResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UserLogoutResponseData"
          }
        },
        "required": [
          "data"
        ]
      },
      "UserSignUpResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UserSignUpResponseData"
          }
        },
        "required": [
          "data"
        ]
      },
      "UserSignUpResponseData": {
        "type": "object",
        "properties": {
          "user": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SafeUser"
              },
              {
                "type": [
                  "object",
                  "null"
                ]
              }
            ]
          },
          "email": {
            "type": "string"
          },
          "email_verification_required": {
            "type": "boolean"
          }
        },
        "required": [
          "user",
          "email",
          "email_verification_required"
        ],
        "additionalProperties": false
      },
      "UserSignUpRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8
          },
          "display_name": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "email",
          "password"
        ],
        "additionalProperties": false
      },
      "UserEmailVerificationConfirmResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UserEmailAckData"
          }
        },
        "required": [
          "data"
        ]
      },
      "UserEmailAckData": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "additionalProperties": false
      },
      "UserEmailVerificationConfirmRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "token": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "email",
          "token"
        ],
        "additionalProperties": false
      },
      "UserPasswordResetRequestResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UserEmailAckData"
          }
        },
        "required": [
          "data"
        ]
      },
      "UserEmailRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          }
        },
        "required": [
          "email"
        ],
        "additionalProperties": false
      },
      "UserPasswordResetConfirmResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UserEmailAckData"
          }
        },
        "required": [
          "data"
        ]
      },
      "UserPasswordResetConfirmRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "token": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "minLength": 8
          }
        },
        "required": [
          "email",
          "token",
          "password"
        ],
        "additionalProperties": false
      },
      "UserResendVerificationResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UserEmailAckData"
          }
        },
        "required": [
          "data"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/v1/user/auth/login": {
      "post": {
        "operationId": "user_auth_login",
        "tags": [
          "user"
        ],
        "x-goldenyears-invalidates": [
          "session",
          "account"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/user_auth_loginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/create_session": {
      "post": {
        "operationId": "create_session",
        "tags": [
          "user"
        ],
        "x-goldenyears-invalidates": [
          "session",
          "account"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/create_sessionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/user/auth/logout": {
      "post": {
        "operationId": "user_auth_logout",
        "tags": [
          "user"
        ],
        "x-goldenyears-invalidates": [
          "session",
          "account"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyJsonBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/user_auth_logoutResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/delete_session": {
      "post": {
        "operationId": "delete_session",
        "tags": [
          "user"
        ],
        "x-goldenyears-invalidates": [
          "session",
          "account"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyJsonBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/delete_sessionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/user/auth/signup": {
      "post": {
        "operationId": "user_auth_signup",
        "tags": [
          "user"
        ],
        "x-goldenyears-invalidates": [
          "session",
          "account"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSignUpRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSignUpResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/user/auth/confirm_verification": {
      "post": {
        "operationId": "user_auth_confirm_verification",
        "tags": [
          "user"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserEmailVerificationConfirmRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserEmailVerificationConfirmResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/user/auth/request_password_reset": {
      "post": {
        "operationId": "user_auth_request_password_reset",
        "tags": [
          "user"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserEmailRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPasswordResetRequestResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/user/auth/confirm_password_reset": {
      "post": {
        "operationId": "user_auth_confirm_password_reset",
        "tags": [
          "user"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserPasswordResetConfirmRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPasswordResetConfirmResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/user/auth/resend_verification": {
      "post": {
        "operationId": "user_auth_resend_verification",
        "tags": [
          "user"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserEmailRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResendVerificationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}